requirements.txt
torch
transformers
opencv-python
numpy
llava-torch
fastapi
uvicorn
python-multipart
pillow